commit 'API Rest CUA AD'

+++ CUA-JGG2425/Flights_JGG/src/main/java/com/jgg/catchup/flights_jgg/models/dao/IContainDAO.java
@@ -1,8 +1,9 @@
 package com.jgg.catchup.flights_jgg.models.dao;
 
+import com.jgg.catchup.flights_jgg.models.entities.Contain;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IContainDAO extends CrudRepository<Contain, Long> {
+public interface IContainDAO extends CrudRepository<Contain, Integer> {
 }
