+++ CUA-JGG2425/Flights_JGG/src/main/java/com/jgg/catchup/flights_jgg/models/dao/ITicketDAO.java
@@ -1,8 +1,9 @@
 package com.jgg.catchup.flights_jgg.models.dao;
 
+import com.jgg.catchup.flights_jgg.models.entities.Ticket;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITicketDAO extends CrudRepository<Ticket, Long> {
+public interface ITicketDAO extends CrudRepository<Ticket, Integer> {
 }
